Make ce-oem-info jobs be attachment jobs
ce-oem-info jobs are aimed to collect the system bootup information from
snapd and cloud-init to help developers analyzing bugs. Therefore, these
jobs are more suitable to be attachement jobs and should not fail even
if some information can not be found. For example, some system may not have
cloud-init. In these system, the job cannot gather the information from
cloud-init, but this should not be a failure.