Revamp caam_hwrng_test by looping w/ smaller size
Some of the project devices' caam_hwrng have buffer, so previously the
dd size is increased to 10M to exceed the buffer size to make the
interrupt increased. The it will make the device without buffer take
very long time (over 30 mins) to do the dd command.

Therefore, this commit separeted the size to 512K and loop it 20 times.
On the device without buffer it could do 512K and detect the increased
interrupt to get the pass result, which is much faster.
On the device with buttfer it could also exceed the buffer by dd 20
times and detect the increased interrupt to get the pass result.